+++ src/components/organisms/MovieDetails/useMovieDetails.tsx
@@ -2,8 +2,13 @@ import { useEffect, useState } from 'react';
 import { ListType } from '../../../utils/types';
 import { useAppSelector } from '../../../app/reduxHooks';
 import { useMovieService } from '../../../hooks/api/useMovieService';
+import { usePagination } from '../../../hooks/usePagination';
 
 export const useMovieDetails = () => {
+  const { currentPage, setCurrentPage, limit, handlePageChange } =
+    usePagination();
+
+  const { isFetching } = useMovieService({ page: currentPage, limit });
   const {
     name,
     description,
@@ -15,16 +20,17 @@ export const useMovieDetails = () => {
     reviews,
   } = useAppSelector(state => state.movie);
 
-  const [currentPage, setCurrentPage] = useState(1);
-  const [limit, setLimit] = useState(10);
-
-  const { isFetching } = useMovieService({ page: currentPage, limit });
-
   const totalRating = rating.imdb !== 0 ? rating.imdb / 2 : rating.kp / 2;
 
   const [selectedValue, setSelectedValue] = useState<ListType>('actors');
-
   const [totalData, setTotalData] = useState([]);
+
+  const handleChangeDetails = (value: ListType) => {
+    setSelectedValue(value);
+  };
+  const handleClick = (id: number) => {
+    console.log('hello', id);
+  };
 
   useEffect(() => {
     setCurrentPage(1);
@@ -38,23 +44,6 @@ export const useMovieDetails = () => {
       setTotalData(reviews);
     }
   }, [selectedValue, persons, seasonsInfo, reviews]);
-
-  useEffect(() => {
-    console.log('DATA', totalData);
-  }, [totalData]);
-
-  const handlePageChange = (newPage: number, newPageSize: number) => {
-    setCurrentPage(newPage);
-    setLimit(newPageSize);
-  };
-
-  const handleChangeDetails = (value: ListType) => {
-    setSelectedValue(value);
-  };
-
-  const handleClick = (id: number) => {
-    console.log('dasdas', id);
-  };
 
   return {
     isFetching,
